Request MEDIA_LOCATION permission
This allows us to get gps coordinates in exif data.

This permission is weird. It doesn't show up as a permission is the
app's settings permission list. It also doesn't have its own prompt
to allow it when you request it. But you still need to ask for it
at runtime for some reason.